Refactoring initialisation de l'éditeur
